Move tests to separate `tests` folder
