+++ tests/analyzer/test_base.py
@@ -1,19 +1,15 @@
 # Get Python six functionality:
-from __future__ import absolute_import, division, print_function, unicode_literals
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
 
 import pytest
 
-from innvestigate.analyzer import BaselineGradient, Gradient
-from innvestigate.utils.tests import dryrun
+from innvestigate.analyzer import BaselineGradient
+from innvestigate.analyzer import Gradient
 
-###############################################################################
-###############################################################################
-###############################################################################
-
-
-###############################################################################
-###############################################################################
-###############################################################################
+from tests import dryrun
 
 
 @pytest.mark.fast
