+++ tests/analyzer/test_init.py
@@ -8,15 +8,6 @@ import pytest
 from innvestigate import create_analyzer
 from innvestigate.analyzer import analyzers
 
-###############################################################################
-###############################################################################
-###############################################################################
-
-
-###############################################################################
-###############################################################################
-###############################################################################
-
 
 @pytest.mark.fast
 @pytest.mark.precommit
