+++ tests/analyzer/test_pattern_based.py
@@ -1,21 +1,17 @@
 # Get Python six functionality:
-from __future__ import absolute_import, division, print_function, unicode_literals
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
 
 import pytest
 
-from innvestigate.analyzer import PatternAttribution, PatternNet
-from innvestigate.utils.tests import dryrun
+from innvestigate.analyzer import PatternAttribution
+from innvestigate.analyzer import PatternNet
 
-###############################################################################
-###############################################################################
-###############################################################################
-
+from tests import dryrun
 
 # todo: add again a traint/test case for mnist
-
-###############################################################################
-###############################################################################
-###############################################################################
 
 
 @pytest.mark.fast
