+++ tests/layer.py
@@ -6,21 +6,11 @@ from builtins import range
 import keras.engine.topology
 import keras.models
 
-from ... import utils as iutils
-
-###############################################################################
-###############################################################################
-###############################################################################
-
+from innvestigate import utils as iutils
 
 __all__ = [
     "TestAnalysisHelper",
 ]
-
-
-###############################################################################
-###############################################################################
-###############################################################################
 
 
 class TestAnalysisHelper(object):
