+++ tests/networks/__init__.py
@@ -1,11 +1,17 @@
 # Get Python six functionality:
-from __future__ import absolute_import, division, print_function, unicode_literals
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
 
 import fnmatch
 
 import keras.backend as K
 
-from . import cifar10, imagenet, mnist, trivia
+from tests.networks import cifar10
+from tests.networks import imagenet
+from tests.networks import mnist
+from tests.networks import trivia
 
 ###############################################################################
 ###############################################################################
