+++ tests/networks/base.py
@@ -5,11 +5,6 @@ from builtins import range
 
 import keras.layers
 
-###############################################################################
-###############################################################################
-###############################################################################
-
-
 __all__ = [
     "log_reg",
     "mlp_2dense",
@@ -19,11 +14,6 @@ __all__ = [
     "cnn_2convb_3dense",
     "cnn_3convb_3dense",
 ]
-
-
-###############################################################################
-###############################################################################
-###############################################################################
 
 
 # TODO: more consistent nameing
