+++ tests/networks/cifar10.py
@@ -3,12 +3,7 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 import keras.backend as K
 
-from . import base
-
-###############################################################################
-###############################################################################
-###############################################################################
-
+from tests.networks import base
 
 __all__ = [
     "log_reg",
@@ -21,11 +16,6 @@ __all__ = [
 ]
 
 
-###############################################################################
-###############################################################################
-###############################################################################
-
-
 if K.image_data_format() == "channels_first":
     __input_shape__ = [None, 3, 32, 32]
 else:
@@ -34,16 +24,12 @@ __output_n__ = 10
 
 
 ###############################################################################
-###############################################################################
-###############################################################################
 
 
 def log_reg(activation=None):
     return base.log_reg(__input_shape__, __output_n__, activation=activation)
 
 
-###############################################################################
-###############################################################################
 ###############################################################################
 
 
@@ -67,8 +53,6 @@ def mlp_3dense(activation=None):
     )
 
 
-###############################################################################
-###############################################################################
 ###############################################################################
