+++ tests/networks/mnist.py
@@ -3,12 +3,7 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 import keras.backend as K
 
-from . import base
-
-###############################################################################
-###############################################################################
-###############################################################################
-
+from tests.networks import base
 
 __all__ = [
     "log_reg",
@@ -34,16 +29,12 @@ __output_n__ = 10
 
 
 ###############################################################################
-###############################################################################
-###############################################################################
 
 
 def log_reg(activation=None):
     return base.log_reg(__input_shape__, __output_n__, activation=activation)
 
 
-###############################################################################
-###############################################################################
 ###############################################################################
 
 
@@ -67,8 +58,6 @@ def mlp_3dense(activation=None):
     )
 
 
-###############################################################################
-###############################################################################
 ###############################################################################
