+++ tests/networks/trivia.py
@@ -3,23 +3,13 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 import keras.layers
 
-from . import base
-
-###############################################################################
-###############################################################################
-###############################################################################
-
+from tests.networks import base
 
 __all__ = [
     "dot",
     # todo: check why this makes problems to wrapper implementation.
     "skip_connection",
 ]
-
-
-###############################################################################
-###############################################################################
-###############################################################################
 
 
 def dot():
