+++ tests/tools/test_pattern.py
@@ -13,16 +13,10 @@ from keras.models import Model
 
 import innvestigate
 from innvestigate.tools import PatternComputer
-from innvestigate.utils.tests import dryrun
 
-###############################################################################
-###############################################################################
-###############################################################################
-
-
-###############################################################################
-###############################################################################
-###############################################################################
+from tests import dryrun
+from tests.networks.base import cnn_2convb_2dense
+from tests.networks.base import mlp_2dense
 
 
 @pytest.mark.fast
@@ -199,7 +193,7 @@ def train_model(model, data, epochs=20):
 class MnistPatternExample_dense_linear(unittest.TestCase):
     def test(self):
         np.random.seed(234354346)
-        model_class = innvestigate.utils.tests.networks.base.mlp_2dense
+        model_class = mlp_2dense
 
         data = fetch_data()
         model, modelp = create_model(model_class)
@@ -240,7 +234,7 @@ class MnistPatternExample_dense_linear(unittest.TestCase):
 class MnistPatternExample_dense_relu(unittest.TestCase):
     def test(self):
         np.random.seed(234354346)
-        model_class = innvestigate.utils.tests.networks.base.mlp_2dense
+        model_class = mlp_2dense
 
         data = fetch_data()
         model, modelp = create_model(model_class)
@@ -336,7 +330,7 @@ class MnistPatternExample_dense_relu(unittest.TestCase):
 #     def test(self):
 #         np.random.seed(234354346)
 #         K.set_image_data_format("channels_first")
-#         model_class = innvestigate.utils.tests.networks.base.cnn_2convb_2dense
+#         model_class = cnn_2convb_2dense
 #         data = fetch_data()
 #         model, modelp = create_model(model_class)
 #         train_model(modelp, data, epochs=1)
