+++ tests/utils/keras/test_graph.py
@@ -5,16 +5,10 @@ import keras.models
 import pytest
 
 from innvestigate.utils.keras import graph as kgraph
-from innvestigate.utils.tests import networks
 
-###############################################################################
-###############################################################################
-###############################################################################
+from tests import networks
 
-
-###############################################################################
-###############################################################################
-###############################################################################
+#############################################################################
 
 
 @pytest.mark.fast
